Wire Alembic, narrow remaining excepts, add lint+CI, extract more UI
- README: redirect 12 docs/* references to docs/archive/*
- database.py: replace ensure_lightweight_migrations call with
  apply_alembic_migrations(); on first-time DB (no alembic_version table)
  stamp head, otherwise upgrade head. Legacy function kept for compat
- Narrow 10 broad excepts to OSError/ValueError/RuntimeError/AttributeError
  and add logger.warning/exception in ocr_service, pdf_split_service,
  pdf_text_service, title_crop_service, cad_parse_service, the 3 remaining
  pipeline step loops, and recognizer dxf_loader/entity_extractor/
  layer_extractor/mtext_extractor/text_extractor
- Add pyproject.toml with ruff + mypy config; per-file ignores cover
  SQLAlchemy Mapped["..."] forward refs and database.py deferred imports
- Auto-fix 126 ruff issues across backend/recognizer (imports sort,
  UP017/UP045 modernizations, removed redundant int() cast in confidence,
  prefixed unused unpacked var in field_fusion, switched
  remove_original_dir_if_empty to contextlib.suppress)
- .github/workflows/ci.yml: ruff + npm build + pytest on push/PR
- Extract AppHeader and ProjectsAside from main.tsx (saves ~110 lines,
  3236 -> 2849 -> 2849 with prop-driven children). Further splits need
  state-management refactor (useReducer/zustand) before they're useful
- Add .env.example documenting CADFENXI_ env vars

+++ recognizer/fusion/confidence.py
@@ -11,7 +11,7 @@ def calculate_confidence_score(field_confidences: dict[str, float]) -> int:
     score = 0.0
     for field_name, weight in FIELD_WEIGHTS.items():
         score += float(field_confidences.get(field_name, 0)) * weight
-    return int(round(max(0, min(100, score))))
+    return round(max(0, min(100, score)))
 
 
 def calculate_trust_level(
